Extract cancelBookingModalContent
