+++ Modules/Sources/PointOfSale/Presentation/Bookings/POSBookingDetailView.swift
@@ -98,7 +98,11 @@ struct POSBookingDetailView: View {
         .background(Color.posSurface)
         .navigationBarHidden(true)
         .posModal(item: $cancelModalState) { state in
-            cancelBookingModalContent(for: state)
+            POSCancelBookingModalContent(
+                state: state,
+                booking: booking,
+                cancelModalState: $cancelModalState
+            )
         }
     }
 
@@ -332,74 +336,6 @@ struct POSBookingDetailView: View {
         return "\(start) – \(end)"
     }
 
-    private var formattedCancelDateTime: String {
-        DateFormatter.dateTimeFormatter.string(from: booking.startDate)
-    }
-}
-
-// MARK: - Cancel Booking Flow
-
-private extension POSBookingDetailView {
-    @ViewBuilder
-    func cancelBookingModalContent(for state: CancelBookingModalState) -> some View {
-        switch state {
-        case .confirmation:
-            POSCancelBookingConfirmationView(
-                bookingNumber: booking.id,
-                serviceName: booking.serviceName,
-                formattedDateTime: formattedCancelDateTime,
-                customerName: booking.customerName,
-                isProcessing: false,
-                onClose: { cancelModalState = nil },
-                onConfirm: {
-                    cancelModalState = .processing
-                    Task { @MainActor in
-                        await performCancelBooking()
-                    }
-                },
-                onBack: { cancelModalState = nil }
-            )
-        case .processing:
-            POSCancelBookingConfirmationView(
-                bookingNumber: booking.id,
-                serviceName: booking.serviceName,
-                formattedDateTime: formattedCancelDateTime,
-                customerName: booking.customerName,
-                isProcessing: true,
-                onClose: {},
-                onConfirm: {},
-                onBack: {}
-            )
-        case .success:
-            POSCancelBookingSuccessView(
-                onDone: { cancelModalState = nil },
-                onClose: { cancelModalState = nil }
-            )
-        case .error:
-            POSRefundErrorView(
-                title: Localization.cancelBookingErrorTitle,
-                subtitle: Localization.cancelBookingErrorSubtitle,
-                onRetry: {
-                    cancelModalState = .processing
-                    Task { @MainActor in
-                        await performCancelBooking()
-                    }
-                },
-                onCancel: { cancelModalState = nil },
-                onClose: { cancelModalState = nil }
-            )
-        }
-    }
-
-    @MainActor
-    func performCancelBooking() async {
-        do {
-            try await bookingsModel.bookingsController.cancelBooking(bookingID: booking.id)
-            cancelModalState = .success
-        } catch {
-            cancelModalState = .error
-        }
-    }
 }
 
 // MARK: - Section Card Modifier
@@ -432,13 +368,6 @@ private extension DateFormatter {
     static let timeOnlyFormatter: DateFormatter = {
         let formatter = DateFormatter()
         formatter.dateStyle = .none
-        formatter.timeStyle = .short
-        return formatter
-    }()
-
-    static let dateTimeFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .medium
         formatter.timeStyle = .short
         return formatter
     }()
@@ -586,15 +515,4 @@ private enum Localization {
         comment: "Menu action to cancel a booking from the POS booking detail view."
     )
 
-    static let cancelBookingErrorTitle = NSLocalizedString(
-        "pos.bookingDetailView.cancelBookingError.title",
-        value: "Failed to cancel booking",
-        comment: "Title shown when cancelling a booking from POS has failed."
-    )
-
-    static let cancelBookingErrorSubtitle = NSLocalizedString(
-        "pos.bookingDetailView.cancelBookingError.subtitle",
-        value: "Please try again.",
-        comment: "Subtitle shown when cancelling a booking from POS has failed."
-    )
 }
